fix(web-domains): fix url route provider regexp
ref: #DCE-128
