+++ packages/manager/apps/web-ongoing-operations/src/pages/dashboard/dns/Dns.spec.tsx
@@ -3,10 +3,10 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
-import { useResourcesIcebergV6 } from '@ovh-ux/manager-react-components';
+import { useFeatureAvailability, useResourcesIcebergV6 } from '@ovh-ux/manager-react-components';
 import { dns } from '@/__mocks__/dns';
 import Dns from '@/pages/dashboard/dns/Dns';
-import { taskMeDns } from '@/constants';
+import { allDomFeatureAvailibility, domainFeatureAvailibility, taskMeDns } from '@/constants';
 import { serviceInfo } from '@/__mocks__/serviceInfo';
 import { useGetDomainInformation } from '@/hooks/data/query';
 import { wrapper } from '@/utils/test.provider';
@@ -27,6 +27,9 @@ vi.mock('@/hooks/data/query', () => ({
 
 describe('Dns datagrid', () => {
   it('fetch in a good way using useResourcesIcebergV6', () => {
+    (useFeatureAvailability as jest.Mock).mockReturnValue({
+          data: {[allDomFeatureAvailibility] : true, [domainFeatureAvailibility] : true}
+        }),
     (useResourcesIcebergV6 as jest.Mock).mockReturnValue({
       flattenData: dns,
       isLoading: false,
@@ -54,6 +57,9 @@ describe('Dns datagrid', () => {
   });
 
   it('Display the datagrid element', async () => {
+    (useFeatureAvailability as jest.Mock).mockReturnValue({
+          data: {[allDomFeatureAvailibility] : true, [domainFeatureAvailibility] : true}
+        }),
     (useResourcesIcebergV6 as jest.Mock).mockReturnValue({
       flattenData: dns,
       isLoading: false,
@@ -81,6 +87,9 @@ describe('Dns datagrid', () => {
   });
 
   it('Display the datagrid element but serviceInfo is undefined', async () => {
+    (useFeatureAvailability as jest.Mock).mockReturnValue({
+          data: {[allDomFeatureAvailibility] : true, [domainFeatureAvailibility] : true}
+        }),
     (useResourcesIcebergV6 as jest.Mock).mockReturnValue({
       flattenData: dns,
       isLoading: false,
