+++ packages/manager/apps/web/client/app/domain/general-informations/domain-general-informations.state.js
@@ -36,7 +36,7 @@ const commonResolves = {
   breadcrumb: () => null,
 };
 
-export default /* @ngInject */ ($stateProvider, $urlRouterProvider) => {
+export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('app.domain.product.information', {
     url: '/information',
     views: {
@@ -87,40 +87,6 @@ export default /* @ngInject */ ($stateProvider, $urlRouterProvider) => {
         }),
     },
   });
-
-  $urlRouterProvider.when(
-    /^\/domain\/[^/]+\/information/,
-    /* @ngInject */ (ovhFeatureFlipping, coreURLBuilder, $state, $location) => {
-      ovhFeatureFlipping
-        .checkFeatureAvailability('web:domains:domains')
-        .then((featureAvailability) => {
-          return featureAvailability.isFeatureAvailable('web:domains:domains');
-        })
-        .then((isLegacyFeatureAvailable) => {
-          ovhFeatureFlipping
-            .checkFeatureAvailability('web-domains:domains')
-            .then((featureAvailability) => {
-              return featureAvailability.isFeatureAvailable(
-                'web-domains:domains',
-              );
-            })
-            .then((isFeatureAvailable) => {
-              if (!isLegacyFeatureAvailable && isFeatureAvailable) {
-                const path = $location.path();
-                const productId = path.split('/')[2];
-                const url = coreURLBuilder.buildURL(
-                  'web-domains',
-                  '#/domain/:serviceName/information',
-                  { serviceName: productId },
-                );
-                return window.location.replace(url);
-              }
-              $state.go('app.domain.product.information');
-              return false;
-            });
-        });
-    },
-  );
 
   $stateProvider.state('app.alldom.domain.information', {
     url: '/information',
